requirements.txt
fastapi
uvicorn
streamlit
requests
pyttsx3
speechrecognition
pyaudio
beautifulsoup4
yfinance
google-generativeai
sentence-transformers
numpy
faiss-cpu
